fix: ✨ Fix environment variable loading at build time
This commit updates the project configuration to use dotenv-webpack to load environment variables at build time instead of runtime. It also removes the dotenv package and related configurations, as well as updates the .gitignore file to exclude the .env file.

Additionally, the firebaseConfig.ts file has been updated to use the environment variables for Firebase configuration.